requirements.txt
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.42.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
swagger-spec-validator==2.7.3
pytz==2020.1
whitenoise==5.1.0